chore(tests): move update-service tests to test-update-service file
