+++ tests/tests/test-update-service.rs
@@ -15,23 +15,35 @@
 
 //! Integration tests of Update Service.
 
+use std::error::Error as StdError;
+use std::sync::Arc;
+use std::time::Duration;
+
+use futures_util::io::Cursor;
+use nv_redfish::update_service::MultipartUpdateParameters;
 use nv_redfish::update_service::UpdateService;
+use nv_redfish::Error;
 use nv_redfish::ServiceRoot;
+use nv_redfish_core::DataStream;
 use nv_redfish_core::EntityTypeRef;
+use nv_redfish_core::ModificationResponse;
+use nv_redfish_core::MultipartUpdateRequest;
 use nv_redfish_core::ODataId;
+use nv_redfish_core::OemMultipartPart;
 use nv_redfish_tests::ami_viking_service_root;
 use nv_redfish_tests::Bmc;
 use nv_redfish_tests::Expect;
 use nv_redfish_tests::ODATA_ID;
 use nv_redfish_tests::ODATA_TYPE;
 use serde_json::json;
-use std::error::Error as StdError;
-use std::sync::Arc;
 use tokio::test;
 
 const UPDATE_SERVICE_DATA_TYPE: &str = "#UpdateService.v1_9_0.UpdateService";
 const SW_INVENTORIES_DATA_TYPE: &str = "#SoftwareInventoryCollection.SoftwareInventoryCollection";
 const SW_INVENTORY_DATA_TYPE: &str = "#SoftwareInventory.v1_4_0.SoftwareInventory";
+
+const UPDATE_SERVICE_URI: &str = "/redfish/v1/UpdateService";
+const MULTIPART_URI: &str = "/redfish/v1/UpdateService/update-multipart";
 
 #[test]
 async fn list_dell_fw_inventores() -> Result<(), Box<dyn StdError>> {
@@ -202,4 +214,184 @@ async fn get_update_service(
         }),
     ));
     Ok(service_root.update_service().await?.unwrap())
+}
+
+#[tokio::test]
+async fn uses_multipart_http_push_uri() -> Result<(), Box<dyn StdError>> {
+    let bmc = Arc::new(Bmc::default());
+
+    bmc.expect(Expect::get("/redfish/v1", service_root_json()));
+    bmc.expect(Expect::get(
+        UPDATE_SERVICE_URI,
+        update_service_json(Some(MULTIPART_URI)),
+    ));
+
+    bmc.expect(Expect::multipart_update(
+        MULTIPART_URI,
+        json!({
+            "ForceUpdate": true,
+            "Targets": ["/redfish/v1/Systems/1"]
+        }),
+        "firmware.bin",
+        json!({
+            "@odata.id": "/redfish/v1/TaskService/Tasks/42",
+            "Id": "42"
+        }),
+    ));
+
+    let root = ServiceRoot::new(Arc::clone(&bmc)).await?;
+    let update_service = root
+        .update_service()
+        .await?
+        .ok_or("expected update service")?;
+    let parameters = MultipartUpdateParameters::builder()
+        .with_force_update(true)
+        .with_targets(vec!["/redfish/v1/Systems/1".to_string()])
+        .build();
+
+    let response = update_service
+        .multipart_update_from_reader::<_, _, serde_json::Value>(
+            &parameters,
+            DataStream::new("firmware.bin", Cursor::new(b"firmware".to_vec()))
+                .with_content_length(8),
+            Duration::from_secs(600),
+        )
+        .await?;
+
+    let ModificationResponse::Entity(body) = response else {
+        return Err(String::from("expected entity response").into());
+    };
+
+    assert_eq!(body["@odata.id"], "/redfish/v1/TaskService/Tasks/42");
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn uses_generated_update_parameters_with_oem_parts() -> Result<(), Box<dyn StdError>> {
+    let bmc = Arc::new(Bmc::default());
+
+    bmc.expect(Expect::get("/redfish/v1", service_root_json()));
+    bmc.expect(Expect::get(
+        UPDATE_SERVICE_URI,
+        update_service_json(Some(MULTIPART_URI)),
+    ));
+
+    let oem = nv_redfish::schema::resource::OemUpdate {
+        additional_properties: json!({
+            "Nvidia": {
+                "Mode": "Rms"
+            }
+        }),
+    };
+    let parameters = MultipartUpdateParameters::builder()
+        .with_force_update(true)
+        .with_targets(vec!["/redfish/v1/Systems/1".to_string()])
+        .with_oem(oem)
+        .build();
+
+    let expected_parameters = serde_json::to_value(&parameters)?;
+
+    bmc.expect(Expect::multipart_update_with_oem_parts(
+        MULTIPART_URI,
+        expected_parameters,
+        "firmware.bin",
+        ["OemNvidia"],
+        json!({
+            "@odata.id": "/redfish/v1/TaskService/Tasks/77",
+            "Id": "77"
+        }),
+    ));
+
+    let root = ServiceRoot::new(Arc::clone(&bmc)).await?;
+    let update_service = root
+        .update_service()
+        .await?
+        .ok_or("expected update service")?;
+
+    let request = MultipartUpdateRequest {
+        update_parameters: &parameters,
+        update_stream: DataStream::new("firmware.bin", Cursor::new(b"firmware".to_vec())),
+        oem_parts: vec![OemMultipartPart::new(
+            "OemNvidia",
+            Cursor::new(br#"{"Mode":"Rms"}"#.to_vec()),
+        )?
+        .with_content_type("application/json")],
+        upload_timeout: Duration::from_secs(600),
+    };
+
+    let response = update_service
+        .multipart_update::<_, _, serde_json::Value>(request)
+        .await?;
+
+    let ModificationResponse::Entity(body) = response else {
+        return Err(String::from("expected entity response").into());
+    };
+
+    assert_eq!(body["@odata.id"], "/redfish/v1/TaskService/Tasks/77");
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn requires_multipart_http_push_uri() -> Result<(), Box<dyn StdError>> {
+    let bmc = Arc::new(Bmc::default());
+
+    bmc.expect(Expect::get("/redfish/v1", service_root_json()));
+
+    // Multipart uploads must use the service-provided endpoint, so a
+    // missing MultipartHttpPushUri should fail before any upload is sent.
+    bmc.expect(Expect::get(UPDATE_SERVICE_URI, update_service_json(None)));
+
+    let root = ServiceRoot::new(Arc::clone(&bmc)).await?;
+    let update_service = root
+        .update_service()
+        .await?
+        .ok_or("expected update service")?;
+
+    let result = update_service
+        .multipart_update_from_reader::<_, _, serde_json::Value>(
+            &MultipartUpdateParameters::default(),
+            DataStream::new("firmware.bin", Cursor::new(b"firmware".to_vec()))
+                .with_content_length(8),
+            Duration::from_secs(600),
+        )
+        .await;
+
+    assert!(matches!(
+        result,
+        Err(Error::UpdateServiceMultipartHttpPushUriNotAvailable)
+    ));
+
+    Ok(())
+}
+
+fn service_root_json() -> serde_json::Value {
+    json!({
+        "@odata.id": "/redfish/v1",
+        "Id": "RootService",
+        "Name": "Root Service",
+        "Links": {
+            "Sessions": {
+                "@odata.id": "/redfish/v1/SessionService/Sessions"
+            }
+        },
+        "UpdateService": {
+            "@odata.id": UPDATE_SERVICE_URI
+        }
+    })
+}
+
+fn update_service_json(multipart_uri: Option<&str>) -> serde_json::Value {
+    let mut body = json!({
+        "@odata.id": UPDATE_SERVICE_URI,
+        "Id": "UpdateService",
+        "Name": "Update Service"
+    });
+
+    if let Some(multipart_uri) = multipart_uri {
+        body["MultipartHttpPushUri"] = json!(multipart_uri);
+    }
+
+    body
 }
